requirements.test.txt
beautifulsoup4==4.12.3
pre-commit==3.7.1
pytest==8.2.0
pytest-cov==5.0.0
pytest-homeassistant-custom-component==0.13.135
